refactor FBlogin as an component from the screen;
